+++ src/main/java/org/pkwmtt/repository/ExamRepository.java
@@ -3,5 +3,5 @@ package org.pkwmtt.repository;
 import org.pkwmtt.entity.Exam;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExamRepository extends JpaRepository<Exam, Long> {
+public interface ExamRepository extends JpaRepository<Exam, Integer> {
 }
