+++ src/main/java/org/pkwmtt/repository/ExamTypeRepository.java
@@ -3,5 +3,5 @@ package org.pkwmtt.repository;
 import org.pkwmtt.entity.ExamType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExamTypeRepository extends JpaRepository<ExamType, Long> {
+public interface ExamTypeRepository extends JpaRepository<ExamType, Integer> {
 }
