+++ src/main/java/org/pkwmtt/repository/OTPCodeRepository.java
@@ -3,5 +3,5 @@ package org.pkwmtt.repository;
 import org.pkwmtt.entity.OTPCode;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OTPCodeRepository extends JpaRepository<OTPCode, Long> {
+public interface OTPCodeRepository extends JpaRepository<OTPCode, Integer> {
 }
